Fix: metrics registry
