+++ deeppavlov/core/common/metrics_registry.json
@@ -1,5 +1,6 @@
 {
   "acc": "deeppavlov.metrics.accuracy:round_accuracy",
+  "accuracy_oos": "deeppavlov.metrics.accuracy:accuracy_oos",
   "average__ner_f1__f1_macro__f1": "deeppavlov.metrics.fmeasure:ner_f1__f1_macro__f1",
   "average__roc_auc__roc_auc__ner_f1": "deeppavlov.metrics.fmeasure:roc_auc__roc_auc__ner_f1",
   "bleu": "deeppavlov.metrics.bleu:bleu",
@@ -9,7 +10,6 @@
   "f1_macro": "deeppavlov.metrics.fmeasure:round_f1_macro",
   "f1_weighted": "deeppavlov.metrics.fmeasure:round_f1_weighted",
   "google_bleu": "deeppavlov.metrics.bleu:google_bleu",
-  "joint_accuracy_in_recall_oos": "deeppavlov.metrics.few_shot_metrics:joint_accuracy_in_recall_oos",
   "kbqa_accuracy": "deeppavlov.metrics.accuracy:kbqa_accuracy",
   "log_loss": "deeppavlov.metrics.log_loss:sk_log_loss",
   "matthews_correlation": "deeppavlov.metrics.correlation:matthews_correlation",
@@ -19,7 +19,7 @@
   "multitask_token_accuracy": "deeppavlov.metrics.accuracy:multitask_token_accuracy",
   "ner_f1": "deeppavlov.metrics.fmeasure:ner_f1",
   "ner_token_f1": "deeppavlov.metrics.fmeasure:ner_token_f1",
-  "oos_scores": "deeppavlov.metrics.few_shot_metrics:oos_scores",
+  "oos_scores": "deeppavlov.metrics.fmeasure:oos_scores",
   "pearson_correlation": "deeppavlov.metrics.correlation:pearson_correlation",
   "per_item_bleu": "deeppavlov.metrics.bleu:per_item_bleu",
   "per_item_dialog_accuracy": "deeppavlov.metrics.accuracy:per_item_dialog_accuracy",
@@ -33,8 +33,6 @@
   "rank_response": "deeppavlov.models.ranking.metrics:rank_response",
   "roc_auc": "deeppavlov.metrics.roc_auc_score:roc_auc_score",
   "sets_accuracy": "deeppavlov.metrics.accuracy:sets_accuracy",
-  "accuracy_oos": "deeppavlov.metrics.few_shot_metrics:accuracy",
-  "balanced_accuracy_oos": "deeppavlov.metrics.few_shot_metrics:balanced_accuracy",
   "slots_accuracy": "deeppavlov.metrics.accuracy:slots_accuracy",
   "spearman_correlation": "deeppavlov.metrics.correlation:spearman_correlation",
   "squad_v1_em": "deeppavlov.metrics.squad_metrics:squad_v1_exact_match",
